# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ProtocolSummary {
  tvl: number;
  totalAssets: number;
  pendingDeposits: number;
  claimableDeposits: number;
  pendingRedemptions: number;
  nav: number;
  yieldRate: number;
  risk: string;
  oracleStatus: string;
  middlewareStatus: string;
  firecrawlStatus: string;
  failureMode: boolean;
  lastEvent: string;
}

export interface RwaAsset {
  assetId: string;
  name: string;
  assetType: string;
  nav: number;
  yieldRate: number;
  custodyStatus: string;
  settlementStatus: string;
  riskStatus: string;
  updatedAt: string;
  dataSource: string;
}

export interface RequestStep {
  label: string;
  state: string;
  detail?: string;
}

export interface ProtocolRequest {
  id: string;
  kind: string;
  owner: string;
  amount: number;
  assetId: string;
  status: string;
  createdAt: string;
  steps: RequestStep[];
  message: string;
  claimableAmount: number;
  claimed?: boolean;
}

export type ProtocolRequestInputKind = typeof ProtocolRequestInputKind[keyof typeof ProtocolRequestInputKind];


export const ProtocolRequestInputKind = {
  deposit: 'deposit',
  redeem: 'redeem',
} as const;

export interface ProtocolRequestInput {
  kind: ProtocolRequestInputKind;
  /** @minimum 1 */
  amount: number;
  /** @minLength 1 */
  owner: string;
}

export type ProcessRequestInputMode = typeof ProcessRequestInputMode[keyof typeof ProcessRequestInputMode];


export const ProcessRequestInputMode = {
  valid: 'valid',
  invalid: 'invalid',
} as const;

export interface ProcessRequestInput {
  mode: ProcessRequestInputMode;
}

export interface ProtocolClaim {
  id: string;
  requestId: string;
  owner: string;
  faceValue: number;
  price: number;
  discount: number;
  settlement: string;
  status: string;
  assetId: string;
}

export interface ClaimListingInput {
  /** @minimum 1 */
  price: number;
}

export interface ClaimPurchaseInput {
  /** @minLength 1 */
  buyer: string;
}

export interface FailureModeInput {
  enabled: boolean;
}