# lib/api-zod/src/generated/types/claimListingInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ClaimListingInput {
  /** @minimum 1 */
  price: number;
}