# lib/api-zod/src/generated/types/processRequestInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProcessRequestInputMode } from './processRequestInputMode';

export interface ProcessRequestInput {
  mode: ProcessRequestInputMode;
}