# lib/api-zod/src/generated/types/protocolClaim.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProtocolClaim {
  id: string;
  requestId: string;
  owner: string;
  faceValue: number;
  price: number;
  discount: number;
  settlement: string;
  status: string;
  assetId: string;
}