# lib/api-zod/src/generated/types/protocolRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RequestStep } from './requestStep';

export interface ProtocolRequest {
  id: string;
  kind: string;
  owner: string;
  amount: number;
  assetId: string;
  status: string;
  createdAt: string;
  steps: RequestStep[];
  message: string;
  claimableAmount: number;
  claimed?: boolean;
}