# lib/api-zod/src/generated/types/protocolRequestInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProtocolRequestInputKind } from './protocolRequestInputKind';

export interface ProtocolRequestInput {
  kind: ProtocolRequestInputKind;
  /** @minimum 1 */
  amount: number;
  /** @minLength 1 */
  owner: string;
}