# lib/api-zod/src/generated/types/protocolSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProtocolSummary {
  tvl: number;
  totalAssets: number;
  pendingDeposits: number;
  claimableDeposits: number;
  pendingRedemptions: number;
  nav: number;
  yieldRate: number;
  risk: string;
  oracleStatus: string;
  middlewareStatus: string;
  firecrawlStatus: string;
  failureMode: boolean;
  lastEvent: string;
}