# lib/api-zod/src/generated/types/rwaAsset.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RwaAsset {
  assetId: string;
  name: string;
  assetType: string;
  nav: number;
  yieldRate: number;
  custodyStatus: string;
  settlementStatus: string;
  riskStatus: string;
  updatedAt: string;
  dataSource: string;
}